Initial background package refactoring
