+++ core/src/main/java/org/radargun/stages/cache/background/PrivateLogChecker.java
@@ -1,27 +1,23 @@
 package org.radargun.stages.cache.background;
 
-import java.util.Random;
-
-import org.radargun.stages.helpers.Range;
-
 /**
- * Checker used for {@link PrivateLogValue non-shared log values}
+ * Checker used for {@link org.radargun.stages.cache.background.PrivateLogValue non-shared log values}
  *
  * @author Radim Vansa &lt;rvansa@redhat.com&gt;
  */
 public class PrivateLogChecker extends LogChecker {
 
-   public PrivateLogChecker(int id, Pool logCheckerPool, BackgroundOpsManager manager) {
-      super(manager.getName() + "-Checker-" + id, manager, logCheckerPool);
+   public PrivateLogChecker(int id, LogCheckerPool pool, BackgroundOpsManager manager) {
+      super(manager.getName() + "-Checker-" + id, manager, pool);
    }
 
    @Override
-   protected AbstractStressorRecord newRecord(AbstractStressorRecord record, long operationId, long seed) {
-      return new StressorRecord((StressorRecord) record, operationId, seed);
+   protected StressorRecord newRecord(StressorRecord record, long operationId, long seed) {
+      return new StressorRecord( record, operationId, seed);
    }
 
    @Override
-   protected Object findValue(AbstractStressorRecord record) throws Exception {
+   protected Object findValue(StressorRecord record) throws Exception {
       // We cannot atomically get the two vars - the stressor could move backup to main or back between the two
       // gets. But the chance for doing this 100 times is small enough.
       Object value = null;
@@ -46,7 +42,7 @@ public class PrivateLogChecker extends LogChecker {
    }
 
    @Override
-   protected boolean containsOperation(Object value, AbstractStressorRecord record) {
+   protected boolean containsOperation(Object value, StressorRecord record) {
       if (value == null) {
          return false;
       }
@@ -63,65 +59,6 @@ public class PrivateLogChecker extends LogChecker {
          log.error("Expected value from threadId " + record.getThreadId() + " but found from " + logValue.getThreadId());
       }
       return false;
-   }
-
-   public static class Pool extends LogChecker.Pool {
-
-      public Pool(int numSlaves, int numThreads, long numEntries, long keyIdOffset, BackgroundOpsManager manager) {
-         super(numThreads, numSlaves, manager);
-         int totalThreads = numThreads * numSlaves;
-         for (int threadId = 0; threadId < totalThreads; ++threadId) {
-            Range range = Range.divideRange(numEntries, totalThreads, threadId).shift(keyIdOffset);
-            log.tracef("Record %d has range %s", threadId, range);
-            addNew(new PrivateLogChecker.StressorRecord(threadId, range));
-         }
-         registerListeners(true); // synchronous listeners
-      }
-
-      @Override
-      public void created(Object key, Object value) {
-         log.trace("Created " + key + " -> " + value);
-         modified(key, value);
-      }
-
-      @Override
-      public void updated(Object key, Object value) {
-         log.trace("Updated " + key + " -> " + value);
-         modified(key, value);
-      }
-
-      @Override
-      protected void modified(Object key, Object value) {
-         if (value instanceof PrivateLogValue) {
-            PrivateLogValue logValue = (PrivateLogValue) value;
-            notify(logValue.getThreadId(), logValue.getOperationId(logValue.size() - 1), key);
-         } else super.modified(key, value);
-      }
-   }
-
-   private static class StressorRecord extends AbstractStressorRecord {
-      private final long keyRangeStart;
-      private final long keyRangeSize;
-
-      public StressorRecord(int threadId, Range keyRange) {
-         super(new Random(threadId), threadId);
-         this.keyRangeStart = keyRange.getStart();
-         this.keyRangeSize = keyRange.getSize();
-         next();
-      }
-
-      public StressorRecord(StressorRecord record, long operationId, long seed) {
-         super(seed, record.threadId, operationId);
-         this.keyRangeStart = record.keyRangeStart;
-         this.keyRangeSize = record.keyRangeSize;
-         next();
-      }
-
-      @Override
-      public void next() {
-         currentKeyId = keyRangeStart + (rand.nextLong() & Long.MAX_VALUE) % keyRangeSize;
-         checkFinished(currentOp++);
-      }
    }
 
 }
